Bug #22107047: "MYSQLD --HELP --VERBOSE" CREATES A FILE IN THE DATA DIRECTORY
This patch makes the server skip opening the binlog index file
if starting with --help.

Without this patch, with an empty data directory, the binlog
index file will be created and left in the data directory.
If we then do 'mysqld --initialize' on the same data directory,
it will fail.