Bug #21040199 INCONSISTENT UNIT IN DATAFILE::M_SIZE (PAGES OR BYTES)
Change SysTablespace::parse_units() so that it returns the number
of pages in the file instead of the number of megabytes.
Then delete SysTablespace::normalize_size().

Also, clean up and distinguish the repeated error messages in
SysTablespace::parse_params() and add tests for each kind of
bad entry for --innodb-data-file-path in innodb_create_db.test.

Approved by Marko in rb#10890