BUG#21864838 SIG11 IN FIELD_BLOB::CMP| AT SQL/FIELD.CC Bug#21881155: UNINTIALIZED READ IN KEY_REC_CMP CAUSE CRASH LATER
Problem :
---------
The issue is in generic partition handler that ensures order of
fetched records across partitions. It does a n-way merge with the
records from multiple partitions with the help of a priority queue.
The queue needs to compare records from different partitions based
on key (key_rec_cmp). key_rec_cmp() fails if the records contain
virtual columns.

Solution :
----------
All these comparisons are on the scan keys (virtual columns) from
the secondary index. For secondary index scan if we could always
return the evaluated secondary index key values in the record then
the issues would be resolved.

Reviewed-by: Jimmy Yang <Jimmy.Yang@oracle.com>

RB: 10686