Bug#21974321: WEIGHT_STRING RESULT IS WRONG IF USED IN A VIEW (AS CHAR CLAUSE IS LOST)
In WEIGHT_STRING, string can be casted before calculating its
weight. String is casted to binary with Item_char_typecast. But
the cast info is lost when the casting happends when it is not
cast to binary.
Solution: add virtual function print() to help
Item_func_weight_string store cast info.