Bug#22177076: ASSERT AT M_THD->VARIABLES.OPTION_BITS
This assert was triggered when DDL log recovery for partitioning
was executed at server startup. The assert checks that
@@autocommit=1 mode has been turned off when doing DD updates
as we otherwise don't have control over when commit/rollback
of DD updates happens.

For normal statements this is handled by code in
mysql_execute_command() which temporarily disables @@autocommit=1
when doing DDL statements. However since the DDL log recovery
bypasses this mechanism, autocommit=1 was not disabled.

This patch fixes the problem by explicitly disabling
@@autocommit=1 mode when doing DDL log recovery.