Bug #19601973: FIND_IN_SET()/VALUES(): WITH INVALID COLUMN IN SUBQUERY CRASHES.
Issue:
======
The VALUES() should be set as INSERT_VALUE_ITEM but it was
wrongly treated as FIELD_ITEM. This lead to issue.

SOLUTION:
=========
Set Item type() of class Item_insert_value as
INSERT_VALUE_ITEM.
In trunk and 5.7 bug is not reproducible as fix for
Bug# 19823076 eliminated the scenario where we look for
not yet fixed Item_value_insert. So added missing tests
for the same.