Bug#21875520 Problems with virtual column indexes
The problem here is that Innodb detects a mismatch between the
value of a column in an index versus the value in the "before" buffer
when attempting an update operation.

The problem occurs because virtual generated column values are
calculated by fill_record() when doing an INSERT operation.
However, if the table definition contains DEFAULT or ON UPDATE functions,
these functions are evaluated after fill_record() is called.
So, if a virtual column depends on a column updated by such function,
a data inconsistency occurs.

The minimum solution to this problem is to re-update the virtual columns
if the table has a DEFAULT or ON UPDATE function and has virtual columns.
This means that in a few cases, we update the virtual columns twice, but
removing this inefficiency would require a larger refactoring, with
higher risk.