Bug#22110284: RBR CONVERSION FROM SMALLER MASTER TYPE TO BIGGER SLAVE TYPE IS BROKEN
Problem: Replication from MEDUMTEXT to LONGTEXT would truncate columns
to 2^16-1 bytes, when running with the new data dictionary.

The regression was caused by a tightening of the requirement on the
type argument passed to make_field (via
Create_field::init_for_tmp_table()) which had been done as part of the
new dictionary work. With this tightening the exact blob subtype was
no longer being deuced from pack_length inside make_field().

Since only the pack_length and not the blob subtype is recorded in the
binlog, this resulted in the conversion table on the slave being
created with column type MYSQL_TYPE_BLOB rather than MYSQL_TYPE_MEDIUM_BLOB,
as the pack_length was no longer taken into consideration.

Solution: Use the replicated pack_length to detect the actual BLOB
type when creating the conversion table at the slave.