Bug #21980284 MYSQLPUMP CAUSES INCONSISTENT BACKUP
Problem: Due to parallelism the threads which start the dump process can view
a different states of the database which can cause the dump to be inconsistent.

Fix: To get a consistent backup we lock the server and flush all the tables.
This is done with FLUSH TABLES WITH READ LOCK (FTWRL). This will ensure that
any further connections will view the same state of all the databases which
is ideal state to take backup.