Bug#21691396  ER_GTID_NEXT_TYPE_UNDEFINED_GROUP ON BINLOGLESS SLAVE AFTER IO THREAD RECONNECT
When slave applier thread is applying a partial transaction left in
relay log, we set m_scope_info[Transaction_ctx::SESSION].m_ha_list
on registering binlog engine. So we missed to set
m_scope_info[Transaction_ctx::SESSION].m_ha_list if binlog is
disabled on slave, then failed to clean the partial transaction
context.

To fix the problem, we clean the partial transaction context if the
SERVER_STATUS_IN_TRANS flag of thd->server_status is set, since the
flag must be set when beginning a new transaction regardless of
binlog is enabled or not.