Bug #22155786: GET NEGATIVE FOUND_ROWS() FOR UNION STMT
Issue:
------
The problem can occur under the following conditions:
1) In a query with UNION, a particular query block
   statement with LIMIT ...OFFSET clause generates zero
   rows.
2) Use of FOUND_ROWS() with SQL_CALC_FOUND_ROWS in the
   prior SELECT statement.

In this case, FOUND_ROWS() might return a negative number
of rows.

SOLUTION:
---------
While accumulating the total number of records to be
returned by each SELECT in a UNION, check if the number of
possible rows is less than the offset. In which case, the
number of rows contributed by that SELECT should be taken
as zero.