BUG#22125441 VALGRIND ERROR ON RPL.RPL_4THREADS_DEADLOCK
Problem:

A replication test case started failing on PB2 valgrind without any
change in the server code around the reported issue.

Analysis:

First of all, PB2 probably started reporting the issue because of the
use of a newer version of valgrind or compiler, as the code involved
in the reported issue wasn't changed.

The valgrind report stated that a "conditional jump or move depended on
uninitialized value" while comparing two arrays of chars using memcmp.

The code that initializes one of the arrays used in the comparison is
setting only its first element as zero, letting the remain of the array
uninitialized.

Fix:

The issue was fixed by initializing the whole array with zeros.