Bug #17865492: CRASH IN ITEM_FUNC_MATCH::FIX_INDEX WITH                MATCH AGAINST ( INTERVAL ... )
Issue:
------
This problem occurs when MATCH ... AGAINST is used to
compare a character column and an aggregate function.

SOLUTION:
---------
While setting-up the fulltext search function, fix_index
tries to check for the presence of the supplied fields in
the index. It does not expect a reference item (in this
case Item_aggregate_ref). Changing the code to retrieve the
real_item will avoid this problem.