WL#6204 - InnoDB persistent max value for autoinc columns
This worklog makes use of the framework introduced by WL#7816 to persist
the autoinc counters. After this worklog there would be some behaviour
changes or important points of autoinc counter:

1. Counters of AUTO_INCREMENT should be strictly incremental, as it works
now, when the server is running.

2. Allocated counters won't be reused after the server restart normally,
this must be guaranteed.

3. Allocated counters won't be reused after the server restart from a
crash, this can NOT be guaranteed, since we can't flush the redo logs
immediately everytime.

4. Rollback of transaction won't revert the counter.

5. 'ALTER TABLE' would not change the counter to a smaller value than
current existing max counter in the table.

6. If a counter is updated to some larger value than current next value,
the larger value should be remembered and next counter should start from
it.

The persistence during DDL is not supported very well now, since some
DDL operations are not atomic now. This would be fixed in WL#7743,
WL#7141, and WL#7016.

RB:9138
Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>
Reviewed-by: Marko Makela <marko.makela@oracle.com>