Bug #22142361: ASSERTION FAILED: STRING ITERATOR NOT DEREFERENCABLE IN DD::EAT_TO
There was an error in the loops iterating over characters in a
property string that made the iterator be dereferenced when
pointing at the end of the string. This gave an assert (i.e.
a warning in the log file, not a server failure) on windows
when building with debug, otherwise it was not noticed.

This patch modifies the loops to avoid dereferencing the iterator
when pointing at the end of the string.

For the record, this patch also removes an obsolete function,
completely unrelated to the topic of the bug report.