Bug #17618162 SSL CONNECTION NOT CONSIDERING THE VALUE SPECIFIED BY MYSQL_OPT_READ_TIMEOUT
Description: SSL connection not considering the value
specified by MYSQL_OPT_READ_TIMEOUT. Connect the client
to server with MYSQL_OPT_READ_TIMEOUT lesser
(i.e like 5 seconds) then the mysql_query() execution
time(i.e in mysql_query() have query as "SELECT SLEEP(10)").
For the above scenario we you get "Lost connection to MySQL
server during query" error. And this is happening in non-ssl
case. But in SSL connection case we are getting the query
execution is successful.

Analysis: When we set the MYSQL_OPT_READ_TIMEOUT as '5'
under SSL connection, we call my_net_set_read_timeout()
in mysql_real_connect(i.e CLI_MYSQL_REAL_CONNECT) and
converts the vio->read_timeout to milliseconds and further
we call run_plugin_auth() and further so on to ssl_do(),
there we will make new vio and copy the timeout values to
it. But here we have the timeouts already in milliseconds.
And we are passing those milliseconds(as seconds) to
vio_timeout(). Thus further those will be converted to
incorrect milliseconds which leads to having 15000000
milliseconds instead of 15000 milliseconds in case of SSL
connections. Hence the read_timeout for the query is more
and we are getting the results successfully.

Fix: Since the updated value of vio->read_timeout and
vio->write_timeout value always be in milliseconds, we
are converting the milliseconds value of vio->read_timeout
in vio_reset to seconds and calling vio_timeout().