Bug#21922202 QUANTIFIED COMPARISON READS UNINITIALIZED DATA
1. Single column case.

The first subquery in the repro is always empty due to the "WHERE 0"
clause and is a Item_singlerow_subselect (which is materialized due to
single column and an AND condition) and will detect the emptiness in
Item_in_optimizer::val_int ca. line 2309:

  cache->store(args[0]);
  cache->cache_value();     <--- leads to call of Item_subselect::exec

However, later[1], we call into Item_in_subselect::exec and we modify
the plan a bit:

   if (need_expr_cache && !left_expr_cache &&
        exec_method == EXEC_MATERIALIZATION &&
        init_left_expr_cache())
      DBUG_RETURN(TRUE);

[1]  called from Item_in_optimizer::val_int's lines:

     /* The subquery has to be evaluated */
     (void) item_subs->val_bool_result(); <--------

the interesting part being the call to init_left_expr_cache(). Now, this
method sets up a cache for the (left) subquery for the IN (SOME) query
by going directly at the field selected in the left subquery:

   Cached_item *cur_item_cache= new_Cached_item(unit->thd,
                                                left_expr->element_index(i),
                                                use_result_field);

Just a little further down, we call (ca line 700 in item_subselect.cc)

   test_if_item_cache_changed

which calls down the Cached_item_int all the way to the Item_field to
get a value, but *does not go through the Item_singlerow_subselect*, so
it misses the fact that we know the value is NULL due to the false WHERE
condition.  Also, there has been no data read for the subquery (empty),
so the field buffer has not been initialized and we see the read of
uninitialized data.

The fix checks if we saw the empty single row result set when first
calling value->cache() in Item_in_optimizer::val_int at the top level,
so that we omit setting up the (broken) left subquery cache in
Item_in_subselect::exec.

This fixes the problem, because this forces the logic to go through
Item_singlerow_subselect which knows the set is empty. We will
materialize the table with an empty set once; getting subsequent rows
through Item_in_optimizer::val_int will see that all left columns are
null, ca line 2368:

    if (all_left_cols_null && result_for_null_param != UNKNOWN &&
        !item_subs->dependent_before_in2exists())
    {
      /*
         This subquery was originally not correlated. The IN->EXISTS
         transformation may have made it correlated, but only to the
         left expression. All values in the left expression are NULL,
         and we have already evaluated the subquery for all NULL values:
         return the same result we did last time without evaluating the
         subquery.
      */
      null_value= result_for_null_param;
    }

and this time result_for_null_param is no longer UNKNOWN, so we do an
early return instead of evaluating the (left) subquery again.

2. Multiple column case

The multiple column case has another code path since it doesn't
materialize, but uses the IN -> EXISTS transformation instead. As we
shall see, this error also pertains to "missing" the information that
the left operand of the IN is a subquery.

[ Update: While work was happening on this patch, another patch went int
  that made the change described within the square brackets here
  redundant.

  e9f7689b179b0f6d773361240f10a5ca33fe9339
  Bug#22131961 CANNOT GET GEOMETRY OBJECT FROM DATA YOU SEND TO THE GEOMETRY FIELD

  In appendix I below I explain why this change made the change
  described here redundant.

  This code had another bug causing an attempt to evaluate the left
  subquery's columns at optimize time [cf. the val_int call in
  make_join_select attempted because the make_cond_for_table method
  mistakenly returned non-NULL due to wrong const-ness, see below], with
  an un-opened table, causing reading of uninitialized data.

  The root cause is that the marking of the condition with
  "depended_from" failed (for the multiple column case; whereas for a
  forced IN -> EXISTS transformation instead of materialization in the
  single column case, there was no issue) in
  Item_in_subselect::row_value_in_to_exists_transformer because it used
  the expressions directly to determine const-ness instead of (also)
  consulting the top level, i.e. the subquery. This is fixed by the new
  method is_really_constant. ]

But there was another issue at execution time for the multiple column
case: The row cache example holds the subquery, and the evaluation of it
during Item_cache_row::cache_value's call to bring_value(), will find
that the subquery contains no rows due to the impossible WHERE and short
circuit evaluation.

In the loop over columns in Item_cache_row::cache_value, we should not
try to cache based on it qua example, since there will have been no rows
read, so the bottom Item_field::val_xxx will read uninitialized
data. Instead, we force the cache value to a null_value using the new
method Item_cache::store_null.

Regression tests have been added.

Appendix I. Why the fix to Bug#22131961 made the introduction of
is_really_constant() redundant.

This patch, among other changes also contains this change (for
Item_cache::Setup):

// --------------------------------- sql/item.h --------------------
// index 10d43fc..916718f 100644
// @@ -5110,6 +5110,10 @@ public:
       if (((Item_field *)item)->table_ref)
         used_table_map= ((Item_field *)item)->table_ref->map();
     }
+    else
+    {
+      used_table_map= item->used_tables();
+    }
     return 0;
   };
   enum Type type() const { return CACHE_ITEM; }

Originally, the test was (before we introduced is_really_constant):

  expr->element_index(i)->const_item()

and in this case, the const_item() used looks like this:

  /*
    Returns true if this is constant (during query execution, i.e. its
    value will not change until next fix_fields) and its value is known.
    When the default implementation of used_tables() is effective, this
    function will always return true (because used_tables() is empty).
  */
  virtual bool const_item() const
  {
    if (used_tables() == 0)
      return can_be_evaluated_now();
    return false;
  }

It was used_tables() that gave 0, i.e. the result was evaluated as
can_be_evaluated_now() which gave true. After the change, used_tables is
not 0, and we return false, which is correct for "a AND 1", cf the query

  SELECT (SELECT a AND 1, a AND 1 FROM t1 WHERE 0) IN (SELECT 1,1 FROM t1)
      FROM t1;

It seems correct that used_tables should return something since a
references t1.