Bug#22224313 ADD -DWITH_BOOST=SYSTEM OPTION AND SPLIT SHIPPED BOOST FILES
The server uses boost files for two different things:

 a) GIS subsystem using boost header files only
 b) mysqlpump building a private version a boost crono, thread and more

A very valid request from the Linux distros is to be able to build against
system installed boost which is maintained and where fixes
are applied in a timely manner.

So, on a build host with the correct version of boost installed
(currently 1.59) this patch adds support for running cmake without
any WITH_BOOST or BOOST_ROOT argument. For uniformity with other
-DWITH_xxx arguments we have, -DWITH_BOOST=system is also supported.