WL#8896: Enable use of C++11 in MySQL server
This patch turns on C++11 support for MySQL 5.8. This is done by
explicitly setting -std=c++11 for GCC, Clang and Solaris Studio.

In order to get a near complete level of C++11 support, the patch
raises the minimum required version of different compilers to:
- GCC 4.8
- Clang 3.4
- Solaris Studio 12.4

Visual Studio minimum is still 2013. It will be raised to 2015
in a separate patch. This will bring C++11 support on Windows
more in line with other compilers.

A consequence of using -std=c++11 for Solaris Studio is that
we no longer link with stlport. Solaris Studio will automatically
use the g++ runtime library instead.

The patch also fixes various new compilation errors and removes
some old obsolete workarounds. Note that new deprecation errors
(std::auto_ptr and the register keyword) are explicitly downgraded
to warnings. It will be up to individual teams to fix these new
warnings.

RPM spec is updated to remove support for RHEL5/OL5 and to
use compilers from Devtoolset to build on RHEL6/OL6. This is
needed to meet new compiler version requirements.

Finally the patch includes some limited testing of new C++11
features (compiler and library).