Bug#22118669 ASSERTION `PFS_FILE != __NULL' FAILED
Before this fix, an assert failed in the server,
in function pfs_end_temp_file_open_wait_and_bind_to_descriptor_v1().

The root cause is that, in the performance schema file instrumentation,
the code should not assume that instrumenting a temporary file open
operation always results in an instrumented file.

In particular, when the server runs out of instrumented file,
a NULL PFS_file pointer is a valid, normal path.

The fix is to relax the assert.