Bug#22130453 HANDLE_FATAL_SIGNAL (SIG=11) AT GET_FILENAME_HASH_PINS
For some combination of consumers,
the performance schema file instrumentation could fail while attempting
to use a NULL pointer, while instrumenting temporary file io.

The root cause is that, in pfs_end_file_close_wait_v1(),
closing a temporary file always uses the thread instrumentation.

This thread instrumentation, however, is only set conditionally,
when the consumer "thread_instrumentation" is enabled.

The fix is to always retreive the thread instrumentation pointer,
when instrumenting file io operation, because it is needed
to maintain the LF_HASH of instrumented files.