Bug#21276561:FAILURE TO GENERATE GTID LEADS TO INCONSISTENCY
Fixing a post push test issue.

POST PUSH TEST ISSUE:
====================

binlog_gtid_exhausted.test has failed in weekly-trunk with
following symptoms.

-ERROR HY000: Binary logging not possible. Message: An error
occurred during flush stage of the commit.
'binlog_error_action' is set to 'ABORT_SERVER'. Hence
aborting the server.
+ERROR HY000: Binary logging not possible. Message: Hence
aborting the server.

Analysis:
========

Existing code uses 'errmsg' char array as both source and
destination within a 'sprintf' statement.

As per the documentation, If copying takes place between
objects that overlap as a result of a call to sprintf() or
snprintf(), the results are undefined.

This might be the reason for the above failure.

Hence added a new destination buffer to hold the resulting
error string.

In addition to that added a new 'master.opt' file with
'--skip-core-file'.