WL#7770: Develop GUNIT test framework and guidelines for DD API
This patch fixes a memory leak detected by ASAN.

The problem was that WL#7770 introduces a fake Object_registry
where the DD objects are stored instead of in real SE. If a gunit
test creates DD objects and does not drop them, then the object
stays in Object_registry dd::Storage_adoptor::m_storage. These
objects are not freed when dd::Storage_adoptor::m_storage
is destroyed (distructor). The solution is to drop the objects
in the test case itself, which avoids the leak.