Bug #21777589 UK IS PARTIALLY DROPPED FROM MYSQLD EVEN WHEN A 'DROP INDEX' FAILS
Unique indexes in cluster have an ordered index internally for scanning
purposes. During a drop index query, the ordered indexes are dropped
first and then the unique indexes are dropped. When the unique index
dropping is rejected for some constraint, the whole query is rejected.
Related index metadata in ha_ndbcluster, m_index[i] gets changed during
the query, but is never restored back to its state when the query is
rejected. Due to this, any subsequent query to that table fails.

This patch fixes that by doing the the unique index removal first and
the ordered index removal later. If unique index removal fails, the
related ordered index removal is skipped totally, and the index metadata
is preserved.