Bug#22268110 VIEW USING HEXADECIMAL OR BIT LITERAL GIVES WRONG RESULTS
Printing of hex/bit literals was limited to the last 8 bytes.
As this code is used for writing the definition
of views and generated columns to the data dictionary,
a view or generated column could produce unwanted results.
The query printed by EXPLAIN (in the warning line) was
also affected.