BUG#22245619 SERVER ABORT AFTER SYNC STAGE OF THE COMMIT FAILS
This is a post-push fix.

Reverted the previous changes and made the binary log sync to use the
MY_IGNORE_BADFD flag when calling mysql_file_sync(). This flag will make
the sync procedure to ignore bad file descriptor errors, so the issue
will not be considered an error anymore.