Bug#22279231 - GET THE INCORRECT AUTO_INCREMENT COLUMN FOR A TABLE WITH VIRTUAL COLUMN
With a table with virtual columns, we will get incorrect column position
for AUTO_INCREMENT column. Because the functions to do the calculation
assumes the passed column position is for non-virtual columns only,
but we use the MySQL column position, which includes the virtual columns.

The fix is to get the proper column position that calculating functions
need according to its original MySQL column position.

RB: 11159
Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>