Patch for bug#22135584
  BINLOG CONTAINS SCHEMA EVENTS IN INCORRECT ORDER

Schema events were appended to the binlog in incorrect order
relative to non-schema events.

Root cause was that the binlog injector did not properly
handle the case where schema events and non-schema events
were from different epochs.

This patch modifies the handling of events from the two
different schema and non-schema event stream such that
events are are handled one epoch at a time - starting with
events from the lowest available epoch.

Furthermore it clearifies the usage of the local 'gci'
variable: It was used for at least 3 different purposes
which caused confusion, and possible bugs, in the code:

1) Used to hold the 'latest available GCI' when polling
   the 'i_ndb'

2) Used to maintain the gci of the eventOp currently being
   handled (As retrieved from nextEvent()).

3) Used as an argument to isConsistent(Uint64 &) into where
   the GCI of a possible inconsistent epoch is returned.