Bug #22216715 IN-CONSISTENT ERROR MESSAGES NOTICED WITH SSL AND WITHOUT SSL UNDER UCS2 CHARSET
Description: In-consistent error messages noticed when we use
ssl and non-ssl connections with unsupported client
charsets(like "utf32","utf16","ucs2")

Analysis: When the client sends the initial packet for the
connection setup, it will send the client charset number
along with the protocol packet. The server receives and
parses it through parse_client_handshake_packet(). In that
before establishing the SSL connection it will try to
validate the client charset and notices that the client
charset is not supported and returns the error. But since
the connection request is sent by the ssl client it will
receive the charset error message and parses it in his
own way and gives the error as
"protocol version mismatch[2026]".

Fix: Moved the client charset validation function
init_client_charset(), after the ssl connection handshake.