Follow up patch for bug#22135584
Fix a regression introduced by prev. batch for above bug:

When calculating the current_epoch to handle events from,
we didn't take the 'no-binlogging' case into considderation.

When there are no binlogging, the 'i_epoch'
(and ndb_latest_received_binlog_epoc) is '0'. This caused
'current_epoch' to always be calculated as '0' as well, effectively
causing no (schema-)events to ever be handled by the BI-thread.