Fixed bug#21980430: GCOLS: CRASHING
DS-MRR works in the following way: 1) Use a secondary handler to scan the
index to collect the row_ids. 2) Sort the row_ids in a buffer. 3) Use the
original handler to scan the rows according to the sorted row_ids. In step
1), the virtual generated columns will be updated if they exist in read_set
because step 1) uses a covering index to collect the row_ids but ignores to
change the read_set. Because the base columns of generated columns are not
filled in step 1), the valgrind warnings are thrown during evaluate the
virtual generated columns.

There are two possible solutions: 1) Set a temporary read_set for the
secondary index scan. In the temporary read_set, we only need set the
row_id column. After DS-MRR is done, we can restore the read_set.
2) Set table->key_read to be TRUE when doing the secondary scan and
restore it after having read the index using the second handler. If
table->key_read is set to be TRUE, during evaluating virtual generated
columns, the server finds it's a covering index scan and the evaluation
of virtual generated columns will be skipped (Details can be referenced
in function update_generated_read_fields).

This patch chooses 2) because it's simpler than 1).